requirements.txt
apache-airflow==2.10.5
apache-airflow-providers-amazon
apache-airflow-providers-microsoft-mssql
boto3
pandas
pyarrow